reconfigure routing, reorganize files
