+++ app/client/App.js
@@ -8,7 +8,6 @@ export default class App extends React.Component {
   render() {
     return (
       <div>
-        <h1>Socialite (hi from App)</h1>
         {this.props.children}
       </div>
     )
