+++ webpack.config.js
@@ -9,7 +9,7 @@ module.exports = {
   devtool: 'eval-source-map',
   entry: [
     'webpack-hot-middleware/client?reload=true',
-    path.join(__dirname, 'app/client/main.js')
+    path.join(__dirname, 'app/client/router.jsx')
   ],
   output: {
     path: path.join(__dirname, '/dist/'),
@@ -19,7 +19,9 @@ module.exports = {
   plugins: [
     new CopyWebpackPlugin([
       {from: 'app/assets', to: 'assets' }
-    ]),
+    ], {
+      ignore: ['**/src/**', '*.txt', '**/scss/**', '**/less/**']
+    }),
     new HtmlWebpackPlugin({
       template: 'app/client/index.tpl.html',
       inject: 'body',
