+++ webpack.production.config.js
@@ -10,7 +10,7 @@ module.exports = {
   devtool: 'eval-source-map',
   entry: [
     'webpack-hot-middleware/client?reload=true',
-    path.join(__dirname, 'app/client/main.js')
+    path.join(__dirname, 'app/client/router.jsx')
   ],
   output: {
     path: path.join(__dirname, '/dist/'),
